WIP make this code more readable
